django - largely implement inital desired rest endpoints
restframeworks autodoc generation seems to work fine for defaultrouter
routes but when I've added PhraseList, because it's not a viewset,
the apiroot doesn't seem to know the endpoint exists.

Added redoc/swagger, which can generate schema with openapi which
seems fine. But it would be nice to workout how to update api root.